+++ src/TextInput.js
@@ -29,6 +29,7 @@ class TextInput extends Component {
 
   render() {
     const {
+      children,
       s,
       m,
       l,
@@ -84,7 +85,10 @@ class TextInput extends Component {
     const renderLabel = () =>
       label && (
         <label
-          className={cx({ active: value || placeholder })}
+          className={cx({
+            active: value || placeholder,
+            'label-icon': typeof label !== 'string'
+          })}
           data-success={success}
           data-error={error}
           htmlFor={inputProps.id}
@@ -102,8 +106,15 @@ class TextInput extends Component {
         />
       );
 
-    const renderIcon = () =>
-      icon && <i className="material-icons prefix">{icon}</i>;
+    const renderIcon = () => {
+      if (!icon) return;
+
+      if (typeof icon === 'string') {
+        return <i className="material-icons prefix">{icon}</i>;
+      }
+
+      return React.cloneElement(icon, { className: 'prefix' });
+    };
 
     return (
       <div className={wrapperClasses}>
@@ -117,12 +128,14 @@ class TextInput extends Component {
         />
         {renderLabel()}
         {renderHelper()}
+        {children}
       </div>
     );
   }
 }
 
 TextInput.propTypes = {
+  children: PropTypes.node,
   /*
    * Strip away all layout classes such as col and sX
    */
@@ -157,13 +170,13 @@ TextInput.propTypes = {
    */
   id: PropTypes.string,
   /*
-   * prefix icon
+   * prefix icon, name of the icon or a node
    */
-  icon: PropTypes.string,
+  icon: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
   /*
    * label text
    */
-  label: PropTypes.string,
+  label: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
   /*
    * Input initial value
    */